+++ backend/canisters/community/impl/src/jobs/mod.rs
@@ -1,6 +1,6 @@
-pub mod import_groups;
-
 use crate::RuntimeState;
+
+pub mod import_groups;
 
 pub(crate) fn start(state: &RuntimeState) {
     import_groups::start_job_if_required(state);