+++ backend/external_canisters/icrc1_ledger/c2c_client/src/lib.rs
@@ -2,6 +2,7 @@ use canister_client::{generate_candid_c2c_call, generate_candid_c2c_call_no_args
 use icrc1_ledger_canister::*;
 
 // Queries
+generate_candid_c2c_call!(icrc1_balance_of);
 generate_candid_c2c_call_no_args!(icrc1_decimals);
 generate_candid_c2c_call_no_args!(icrc1_fee);
 generate_candid_c2c_call_no_args!(icrc1_metadata);