+++ backend/libraries/types/src/cryptocurrency.rs
@@ -26,7 +26,7 @@ impl Cryptocurrency {
         }
     }
 
-    pub fn decimals(&self) -> Option<usize> {
+    pub fn decimals(&self) -> Option<u8> {
         match self {
             Cryptocurrency::InternetComputer => Some(8),
             Cryptocurrency::SNS1 => Some(8),
@@ -140,6 +140,25 @@ impl PendingCryptoTransaction {
         match self {
             PendingCryptoTransaction::NNS(t) => t.amount.e8s().into(),
             PendingCryptoTransaction::ICRC1(t) => t.amount,
+        }
+    }
+
+    pub fn user_id(&self) -> Option<UserId> {
+        match self {
+            PendingCryptoTransaction::NNS(t) => {
+                if let UserOrAccount::User(u) = t.to {
+                    Some(u)
+                } else {
+                    None
+                }
+            }
+            PendingCryptoTransaction::ICRC1(t) => {
+                if t.to.subaccount.unwrap_or_default() == DEFAULT_SUBACCOUNT.0 {
+                    Some(t.to.owner.into())
+                } else {
+                    None
+                }
+            }
         }
     }