+++ frontend/app/src/components/home/GroupPermissionsEditor.svelte
@@ -51,3 +51,7 @@
     {roles}
     label={$_("permissions.replyInThread")}
     bind:rolePermission={permissions.replyInThread} />
+<!-- <SelectPermissionRole
+    {roles}
+    label={$_("permissions.mentionAllMembers", { values: { mention: "@everyone" } })}
+    bind:rolePermission={permissions.mentionAllMembers} /> -->