+++ frontend/openchat-agent/src/services/common/chatMappers.ts
@@ -30,6 +30,7 @@ import type {
     ApiTotalPollVotes,
     ApiPollConfig,
     ApiGroupPermissions,
+    ApiGroupPermissionsReduced,
     ApiPermissionRole,
     ApiGiphyContent,
     ApiGiphyImageVariant,
@@ -833,6 +834,23 @@ export function groupPermissions(candid: ApiGroupPermissions): ChatPermissions {
         sendMessages: permissionRole(candid.send_messages),
         reactToMessages: permissionRole(candid.react_to_messages),
         replyInThread: permissionRole(candid.reply_in_thread),
+        mentionAllMembers: permissionRole(candid.mention_all_members),
+    };
+}
+
+export function groupPermissionsReduced(candid: ApiGroupPermissionsReduced): ChatPermissions {
+    return {
+        changeRoles: permissionRole(candid.change_roles),
+        updateGroup: permissionRole(candid.update_group),
+        inviteUsers: permissionRole(candid.invite_users),
+        removeMembers: permissionRole(candid.remove_members),
+        deleteMessages: permissionRole(candid.delete_messages),
+        pinMessages: permissionRole(candid.pin_messages),
+        createPolls: permissionRole(candid.create_polls),
+        sendMessages: permissionRole(candid.send_messages),
+        reactToMessages: permissionRole(candid.react_to_messages),
+        replyInThread: permissionRole(candid.reply_in_thread),
+        mentionAllMembers: "admin",
     };
 }
 
@@ -899,6 +917,7 @@ export function apiGroupPermissions(permissions: ChatPermissions): ApiGroupPermi
         react_to_messages: apiPermissionRole(permissions.reactToMessages),
         reply_in_thread: apiPermissionRole(permissions.replyInThread),
         add_members: apiPermissionRole("owner"), // TODO remove this
+        mention_all_members: apiPermissionRole(permissions.mentionAllMembers),
     };
 }