+++ frontend/openchat-agent/src/services/community/candid/idl.js
@@ -135,6 +135,7 @@ export const idlFactory = ({ IDL }) => {
   });
   const GroupPermissions = IDL.Record({
     'block_users' : PermissionRole,
+    'mention_all_members' : PermissionRole,
     'change_permissions' : PermissionRole,
     'delete_messages' : PermissionRole,
     'send_messages' : PermissionRole,
@@ -716,6 +717,7 @@ export const idlFactory = ({ IDL }) => {
     'NameTooShort' : FieldTooShortResult,
     'NotAuthorized' : IDL.Null,
     'Success' : IDL.Record({ 'user_group_id' : IDL.Nat32 }),
+    'NameInvalid' : IDL.Null,
     'UserSuspended' : IDL.Null,
     'CommunityFrozen' : IDL.Null,
     'NameTooLong' : FieldTooLongResult,
@@ -1540,6 +1542,7 @@ export const idlFactory = ({ IDL }) => {
   });
   const OptionalGroupPermissions = IDL.Record({
     'block_users' : IDL.Opt(PermissionRole),
+    'mention_all_members' : IDL.Opt(PermissionRole),
     'change_permissions' : IDL.Opt(PermissionRole),
     'delete_messages' : IDL.Opt(PermissionRole),
     'send_messages' : IDL.Opt(PermissionRole),
@@ -1638,6 +1641,7 @@ export const idlFactory = ({ IDL }) => {
     'NotAuthorized' : IDL.Null,
     'Success' : IDL.Null,
     'UserGroupNotFound' : IDL.Null,
+    'NameInvalid' : IDL.Null,
     'UserSuspended' : IDL.Null,
     'CommunityFrozen' : IDL.Null,
     'NameTooLong' : FieldTooLongResult,