+++ frontend/openchat-agent/src/services/user/candid/types.d.ts
@@ -755,7 +755,7 @@ export interface GroupChatCreated {
 }
 export interface GroupChatSummary {
   'is_public' : boolean,
-  'permissions' : GroupPermissions,
+  'permissions' : GroupPermissionsReduced,
   'metrics' : ChatMetrics,
   'subtype' : [] | [GroupSubtype],
   'date_last_pinned' : [] | [TimestampMillis],
@@ -845,6 +845,22 @@ export interface GroupNameChanged {
   'previous_name' : string,
 }
 export interface GroupPermissions {
+  'block_users' : PermissionRole,
+  'mention_all_members' : PermissionRole,
+  'change_permissions' : PermissionRole,
+  'delete_messages' : PermissionRole,
+  'send_messages' : PermissionRole,
+  'remove_members' : PermissionRole,
+  'update_group' : PermissionRole,
+  'invite_users' : PermissionRole,
+  'change_roles' : PermissionRole,
+  'add_members' : PermissionRole,
+  'create_polls' : PermissionRole,
+  'pin_messages' : PermissionRole,
+  'reply_in_thread' : PermissionRole,
+  'react_to_messages' : PermissionRole,
+}
+export interface GroupPermissionsReduced {
   'block_users' : PermissionRole,
   'change_permissions' : PermissionRole,
   'delete_messages' : PermissionRole,
@@ -1216,6 +1232,7 @@ export interface OptionalCommunityPermissions {
 export interface OptionalContact { 'nickname' : TextUpdate, 'user_id' : UserId }
 export interface OptionalGroupPermissions {
   'block_users' : [] | [PermissionRole],
+  'mention_all_members' : [] | [PermissionRole],
   'change_permissions' : [] | [PermissionRole],
   'delete_messages' : [] | [PermissionRole],
   'send_messages' : [] | [PermissionRole],