+++ backend/canisters/group_index/CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed 
+
+- Accept calls to `set_community_upgrade_concurrency` ([#4418](https://github.com/open-chat-labs/open-chat/pull/4418))
+
+## [[2.0.857](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.857-group_index)] - 2023-09-21
+
 ### Added
 
 - Add `mention_all_members` group permission ([#4405](https://github.com/open-chat-labs/open-chat/pull/4405))